refactor: separate currency enums and models into dedicated files
